Decompile ciMenuTextBioName

+++ src/include/game/data/data_020df0.h
@@ -333,7 +333,7 @@ extern u32 var80088808;
 extern u32 var8008880c;
 extern u32 var80088810;
 extern u32 var80088854;
-extern u32 var800888a0;
+extern u8 var800888a0;
 extern struct chrbio g_ChrBios[10];
 extern u32 misc_bios[];
 extern u32 var80088964;
